+++ travis/perl/before_install.sh
@@ -1,2 +1,0 @@
-#!/bin/bash -x
-source $TRAVIS_BUILD_DIR/travis/error_handler.sh