+++ travis/ruby/script.sh
@@ -1,5 +1,5 @@
 #!/bin/bash -x
 source $TRAVIS_BUILD_DIR/travis/error_handler.sh
-pushd $PROJECT_DIR
-bundle exec rake spec:all
+pushd $TRAVIS_BUILD_DIR/src/ruby
+rake test:all
 popd